create hooks for fetching pref lists
