+++ src/api/responseType.ts
@@ -1,0 +1,6 @@
+import { Prefecture } from '@commonType';
+
+export type APIResponsePrefLists = {
+  message: string | null;
+  result: Prefecture[];
+};
