+++ src/commonType.ts
@@ -1,0 +1,4 @@
+export type Prefecture = {
+  prefCode: number;
+  prefName: string;
+};
